Move DID from pkg/crypto/did to pkg/did
