+++ pkg/crypto/jwt/jwt_test.go
@@ -7,8 +7,8 @@ import (
 	"strings"
 	"testing"
 
-	"github.com/sonr-io/sonr/pkg/crypto/did"
-	"github.com/sonr-io/sonr/pkg/crypto/did/ssi"
+	"github.com/sonr-io/sonr/pkg/did"
+	"github.com/sonr-io/sonr/pkg/did/ssi"
 	"github.com/stretchr/testify/assert"
 	"github.com/tendermint/tendermint/crypto/secp256k1"
 )
