+++ pkg/did/credential.go
