+++ pkg/did/did_test.go
