+++ pkg/did/interface.go
@@ -3,7 +3,7 @@ package did
 import (
 	"encoding/json"
 
-	"github.com/sonr-io/sonr/pkg/crypto/did/ssi"
+	"github.com/sonr-io/sonr/pkg/did/ssi"
 )
 
 type Document interface {
