+++ pkg/did/internal/marshal/marshal.go
