+++ pkg/did/internal/marshal/marshal_test.go
