+++ pkg/did/ssi/spec-registries.go
