+++ pkg/did/ssi/types.go
