+++ pkg/did/ssi/types_test.go
