+++ pkg/did/utils.go
