+++ pkg/did/validator.go
