+++ pkg/motor/motor.go
@@ -9,9 +9,9 @@ import (
 	"github.com/sonr-io/multi-party-sig/pkg/party"
 	"github.com/sonr-io/sonr/pkg/client"
 	"github.com/sonr-io/sonr/pkg/config"
-	"github.com/sonr-io/sonr/pkg/crypto/did"
-	"github.com/sonr-io/sonr/pkg/crypto/did/ssi"
 	"github.com/sonr-io/sonr/pkg/crypto/mpc"
+	"github.com/sonr-io/sonr/pkg/did"
+	"github.com/sonr-io/sonr/pkg/did/ssi"
 	"github.com/sonr-io/sonr/pkg/host"
 	st "github.com/sonr-io/sonr/x/schema/types"
 	rtmv1 "go.buf.build/grpc/go/sonr-io/motor/api/v1"
