+++ pkg/vault/create.go
@@ -9,8 +9,8 @@ import (
 	"net/http"
 	"time"
 
-	"github.com/sonr-io/sonr/pkg/crypto/did"
-	"github.com/sonr-io/sonr/pkg/crypto/did/ssi"
+	"github.com/sonr-io/sonr/pkg/did"
+	"github.com/sonr-io/sonr/pkg/did/ssi"
 )
 
 var (
