+++ pkg/vault/issue-shard.go
@@ -3,7 +3,7 @@ package vault
 import (
 	"errors"
 
-	"github.com/sonr-io/sonr/pkg/crypto/did"
+	"github.com/sonr-io/sonr/pkg/did"
 )
 
 func (v *vaultImpl) IssueShard(shardPrefix, dscPub, dscShard string) (did.Service, error) {
