+++ x/registry/types/did.go
@@ -4,7 +4,7 @@ import (
 	"encoding/json"
 	"fmt"
 
-	"github.com/sonr-io/sonr/pkg/crypto/did"
+	"github.com/sonr-io/sonr/pkg/did"
 )
 
 const contextKey = "@context"
