+++ x/schema/simulation/simap.go
@@ -9,8 +9,8 @@ import (
 
 	sdk "github.com/cosmos/cosmos-sdk/types"
 	simtypes "github.com/cosmos/cosmos-sdk/types/simulation"
-	"github.com/sonr-io/sonr/pkg/crypto/did"
-	"github.com/sonr-io/sonr/pkg/crypto/did/ssi"
+	"github.com/sonr-io/sonr/pkg/did"
+	"github.com/sonr-io/sonr/pkg/did/ssi"
 	"github.com/sonr-io/sonr/x/schema/types"
 )
